job-listings-backend/requirements.txt
azure-cosmos
bs4
Flask
pandas
requests
selenium
webdriver-manager